20230802 version 2.06

+++ accountapp/views.py
@@ -3,6 +3,7 @@ from django.contrib.auth.views import LoginView
 from django.db import transaction
 from django.shortcuts import get_object_or_404, redirect
 from django.urls import reverse_lazy, reverse
+from django.views.decorators.cache import never_cache
 from django.views.generic import CreateView, UpdateView, DeleteView, DetailView, RedirectView
 
 from accountapp.decorators import *
@@ -11,7 +12,7 @@ from accountapp.models import CustomUser
 from phonenumberapp.models import Phonenumber
 from django.utils.decorators import method_decorator
 
-
+@method_decorator(never_cache, name='dispatch')
 class AccountLoginView(LoginView):
     form_class = AccountLoginForm
     template_name = 'accountapp/login.html'
@@ -22,6 +23,7 @@ class AccountLoginView(LoginView):
             return redirect('homescreenapp:homescreen')
         return super().dispatch(request, *args, **kwargs)
 
+@method_decorator(never_cache, name='dispatch')
 @method_decorator(AccountCreateDecorator, name='dispatch')
 class AccountCreateView(CreateView):
     model = CustomUser
@@ -37,12 +39,17 @@ class AccountCreateView(CreateView):
     def form_valid(self, form):
         phonenumber = get_object_or_404(Phonenumber, pk=self.kwargs['pk'])
         phone = phonenumber.phonenumber
+        agree_terms = form.cleaned_data['agree_terms']
         with transaction.atomic():
+            if not agree_terms == True:
+                form.add_error('agree_terms', '약관 및 방침에 동의해 주세요')
+                return self.form_invalid(form)
             form.instance.username= phone
             form.instance.save()
             phonenumber.delete()
             return super().form_valid(form)
 
+@method_decorator(never_cache, name='dispatch')
 @method_decorator(login_required, name='dispatch')
 @method_decorator(AccountOwnerDecorator, name='dispatch')
 class AccountInfoUpdateView(UpdateView):
@@ -54,6 +61,7 @@ class AccountInfoUpdateView(UpdateView):
     def get_success_url(self):
         return reverse_lazy('accountapp:setting', kwargs={'pk': self.object.pk})
 
+@method_decorator(never_cache, name='dispatch')
 @method_decorator(login_required, name='dispatch')
 @method_decorator(AccountOwnerDecorator, name='dispatch')
 class AccountPasswordUpdateView(UpdateView):
@@ -63,6 +71,7 @@ class AccountPasswordUpdateView(UpdateView):
     success_url = reverse_lazy('accountapp:logout')
     template_name = 'accountapp/passwordupdate.html'
 
+@method_decorator(never_cache, name='dispatch')
 @method_decorator(login_required, name='dispatch')
 @method_decorator(AccountOwnerDecorator, name='dispatch')
 class AccountPasswordResetView(UpdateView):
@@ -72,6 +81,7 @@ class AccountPasswordResetView(UpdateView):
     success_url = reverse_lazy('accountapp:logout')
     template_name = 'accountapp/passwordreset.html'
 
+@method_decorator(never_cache, name='dispatch')
 @method_decorator(login_required, name='dispatch')
 @method_decorator(AccountDeleteDecorator, name='delete')
 @method_decorator(AccountGetDeleteDecorator, name='get')
@@ -81,12 +91,14 @@ class AccountDeleteView(DeleteView):
     success_url = reverse_lazy('accountapp:login')
     template_name = 'accountapp/delete.html'
 
+
 @method_decorator(login_required, name='dispatch')
 @method_decorator(AccountOwnerDecorator, name='dispatch')
 class AccountSettingView(DetailView):
     model = CustomUser
     context_object_name = 'target_user'
     template_name = 'accountapp/setting.html'
+
 
 @method_decorator(login_required, name='dispatch')
 @method_decorator(AccountNotificationUpdateDecorator, name='dispatch')
